+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.19.1 (2025-11-6)
+
+- Add `verify_ecdsa_k256_keccak` procedure for verifying signatures using the `miden-crypto` format ([#2344](https://github.com/0xMiden/miden-vm/pull/2344)).
+
 ## 0.19.0 (2025-11-1)
 
 #### Enhancements